clean a little bit the lib compilation still not satisfied

+++ lib/libc/src/errno/errno.c
@@ -1,5 +1,5 @@
-int errno = 0;
+long errno = 0;
 
-int *__get_errno_location(void) {
+long *__get_errno_location(void) {
     return &errno;
 }
